Moving Docker to data access layer
