+++ broker/lib/fabrik/DockerInstance.js
@@ -5,7 +5,7 @@ const Promise = require('bluebird');
 const logger = require('../../../common/logger');
 const errors = require('../../../common/errors');
 const utils = require('../utils');
-const docker = require('../docker');
+const docker = require('../../../data-access-layer/docker');
 const BaseInstance = require('./BaseInstance');
 const catalog = require('../../../common/models').catalog;
 const Timeout = errors.Timeout;
