+++ broker/lib/fabrik/DockerManager.js
@@ -4,7 +4,7 @@ const assert = require('assert');
 const _ = require('lodash');
 const Promise = require('bluebird');
 const config = require('../../../common/config');
-const docker = require('../docker');
+const docker = require('../../../data-access-layer/docker');
 const BaseManager = require('./BaseManager');
 const DockerInstance = require('./DockerInstance');
 const errors = require('../../../common/errors');
