+++ test/test_broker/acceptance/dashboard.docker.spec.js
@@ -5,7 +5,7 @@ const parseUrl = require('url').parse;
 const lib = require('../../../broker/lib');
 const app = require('../support/apps').external;
 const catalog = require('../../../common/models').catalog;
-const docker = lib.docker;
+const docker = require('../../../data-access-layer/docker');
 const fabrik = lib.fabrik;
