+++ test/test_broker/acceptance/service-broker-api.instances.docker.spec.js
@@ -5,7 +5,7 @@ const parseUrl = require('url').parse;
 const lib = require('../../../broker/lib');
 const app = require('../support/apps').internal;
 const catalog = require('../../../common/models').catalog;
-const docker = lib.docker;
+const docker = require('../../../data-access-layer/docker');
 const config = require('../../../common/config');
 const fabrik = lib.fabrik;
