+++ test/test_broker/docker.DockerPortRegistry.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const proxyquire = require('proxyquire');
-const DockerPortRegistry = proxyquire('../../broker/lib/docker/DockerPortRegistry', {
+const DockerPortRegistry = proxyquire('../../data-access-layer/docker/DockerPortRegistry', {
   lodash: {
     sample: function (collection) {
       return collection[2];
